# typechecker+runtime: Perf improvements and some hash salt in your dicts & sets