--- a/bootstrap/stage0/__prelude___specializations.cpp
+++ b/bootstrap/stage0/__prelude___specializations.cpp
@@ -27,6 +27,7 @@
 #include "jakt__platform.h"
 #include "jakt__arguments.h"
 #include "jakt__prelude__prelude.h"
+#include "jakt__prelude__hash.h"
 #include "jakt__prelude__operators.h"
 #include "jakt__prelude__string.h"
 #include "jakt__prelude__iteration.h"