--- a/bootstrap/stage0/error.h
+++ b/bootstrap/stage0/error.h
@@ -3,6 +3,20 @@
 #include "utility.h"
 namespace Jakt {
 namespace error {
+namespace MessageSeverity_Details {
+struct Hint {
+};
+struct Error {
+};
+}
+struct MessageSeverity : public Variant<MessageSeverity_Details::Hint, MessageSeverity_Details::Error> {
+using Variant<MessageSeverity_Details::Hint, MessageSeverity_Details::Error>::Variant;
+    using Hint = MessageSeverity_Details::Hint;
+    using Error = MessageSeverity_Details::Error;
+ErrorOr<DeprecatedString> debug_description() const;
+ErrorOr<DeprecatedString> ansi_color_code() const;
+ErrorOr<DeprecatedString> name() const;
+};
 namespace JaktError_Details {
 struct Message {
 DeprecatedString message;
@@ -34,30 +48,16 @@ using Variant<JaktError_Details::Message, JaktError_Details::MessageWithHint>::V
 ErrorOr<DeprecatedString> debug_description() const;
 utility::Span span() const;
 };
-namespace MessageSeverity_Details {
-struct Hint {
-};
-struct Error {
-};
-}
-struct MessageSeverity : public Variant<MessageSeverity_Details::Hint, MessageSeverity_Details::Error> {
-using Variant<MessageSeverity_Details::Hint, MessageSeverity_Details::Error>::Variant;
-    using Hint = MessageSeverity_Details::Hint;
-    using Error = MessageSeverity_Details::Error;
-ErrorOr<DeprecatedString> debug_description() const;
-ErrorOr<DeprecatedString> ansi_color_code() const;
-ErrorOr<DeprecatedString> name() const;
-};
 }
 } // namespace Jakt
-template<>struct Jakt::Formatter<Jakt::error::JaktError> : Jakt::Formatter<Jakt::StringView>{
-Jakt::ErrorOr<void> format(Jakt::FormatBuilder& builder, Jakt::error::JaktError const& value) {
+template<>struct Jakt::Formatter<Jakt::error::MessageSeverity> : Jakt::Formatter<Jakt::StringView>{
+Jakt::ErrorOr<void> format(Jakt::FormatBuilder& builder, Jakt::error::MessageSeverity const& value) {
 JaktInternal::PrettyPrint::ScopedEnable pretty_print_enable { m_alternative_form };Jakt::ErrorOr<void> format_error = Jakt::Formatter<Jakt::StringView>::format(builder, MUST(value.debug_description()));return format_error;}
 };
 namespace Jakt {
 } // namespace Jakt
-template<>struct Jakt::Formatter<Jakt::error::MessageSeverity> : Jakt::Formatter<Jakt::StringView>{
-Jakt::ErrorOr<void> format(Jakt::FormatBuilder& builder, Jakt::error::MessageSeverity const& value) {
+template<>struct Jakt::Formatter<Jakt::error::JaktError> : Jakt::Formatter<Jakt::StringView>{
+Jakt::ErrorOr<void> format(Jakt::FormatBuilder& builder, Jakt::error::JaktError const& value) {
 JaktInternal::PrettyPrint::ScopedEnable pretty_print_enable { m_alternative_form };Jakt::ErrorOr<void> format_error = Jakt::Formatter<Jakt::StringView>::format(builder, MUST(value.debug_description()));return format_error;}
 };
 namespace Jakt {