--- a/bootstrap/stage0/jakt__prelude__hash.cpp
+++ b/bootstrap/stage0/jakt__prelude__hash.cpp
@@ -1,0 +1,5 @@
+#include "jakt__prelude__hash.h"
+namespace Jakt {
+namespace jakt__prelude__hash {
+}
+} // namespace Jakt