--- a/bootstrap/stage0/jakt__prelude__hash.h
+++ b/bootstrap/stage0/jakt__prelude__hash.h
@@ -1,0 +1,7 @@
+#pragma once
+#include "__unified_forward.h"
+#include <Jakt/Forward.h>
+namespace Jakt {
+namespace jakt__prelude__hash {
+}
+} // namespace Jakt