--- a/bootstrap/stage0/jakt__prelude__prelude.h
+++ b/bootstrap/stage0/jakt__prelude__prelude.h
@@ -3,6 +3,7 @@
 #include "jakt__prelude__iteration.h"
 #include "jakt__prelude__string.h"
 #include "jakt__prelude__operators.h"
+#include "jakt__prelude__hash.h"
 namespace Jakt {
 namespace jakt__prelude__prelude {
 }