--- a/bootstrap/stage0/lexer_specializations.cpp
+++ b/bootstrap/stage0/lexer_specializations.cpp
@@ -28,6 +28,7 @@
 #include "jakt__platform.h"
 #include "jakt__arguments.h"
 #include "jakt__prelude__prelude.h"
+#include "jakt__prelude__hash.h"
 #include "jakt__prelude__operators.h"
 #include "jakt__prelude__string.h"
 #include "jakt__prelude__iteration.h"