… CMakeLists.txt
… bootstrap/stage0/formatter.h
… bootstrap/stage0/lexer.h
… bootstrap/stage0/parser.h
… bootstrap/stage0/repl_backend__common.h
… bootstrap/stage0/runtime/AK/IntrusiveList.h
… bootstrap/stage0/runtime/AK/Variant.h
… bootstrap/stage0/runtime/Builtins/Dictionary.h
… bootstrap/stage0/runtime/Builtins/Set.h
… bootstrap/stage0/runtime/Jakt/Forward.h
… bootstrap/stage0/runtime/jaktlib/prelude/hash.jakt
… bootstrap/stage0/runtime/jaktlib/prelude/prelude.jakt
… bootstrap/stage0/typechecker.cpp
… bootstrap/stage0/typechecker.h
… bootstrap/stage0/typechecker_specializations.cpp
… bootstrap/stage0/types.h
… bootstrap/stage0/types_specializations.cpp
… bootstrap/stage0/utility.h
… bootstrap/stage0/utility_specializations.cpp
… runtime/Builtins/Dictionary.h
… runtime/Builtins/Set.h
… runtime/Jakt/Forward.h
… runtime/jaktlib/prelude/hash.jakt
… runtime/jaktlib/prelude/prelude.jakt
… samples/traits/basic.jakt
… selfhost/codegen.jakt
… selfhost/ids.jakt
… selfhost/interpreter.jakt
… selfhost/typechecker.jakt
… selfhost/types.jakt