Depends on fixes in #1423.

Our friend `for_each_scope_accessible_unqualified_from_scope` _really_ loved to convert ScopeIds into strings to put them in a set, and just in case you're not certain, that is indeed a Very Bad Thing to do if you care about performance at all.
This PR introduces the Hashable trait, which allows the user to place any random object implementing it in a set or use it as a dict key.
Using the magical method of wonders called "Strings No Mo", this PR makes the compiler go (a little) faster:

| | Before | After | How much faster is that? |
| :-- | :-: | :-: | :- | 
| Deboog build | 43.5s | 18.2s | 58% |
| Release build | 2.6s | 1.4s | 46% |

tldr:
```
struct Foo implements(Hashable, Equal<Foo>) {
    x: Stuff1
    y: Stuff2
    
    fn equals(this, anon rhs: Foo) -> bool => .x == rhs.x and .y == rhs.y
    fn hash(this) -> u64 => .x.pair_hash_with(&.y)
}

fn foo() {
    let set = {Foo(x: stuff1, y: stuff2)} // yay
}
```